Add UseExisting provisioning + happy path test
