+++ crates/holochain/src/conductor/interface/websocket.rs
@@ -564,6 +564,7 @@ pub mod test {
             agent_key: agent_key.clone(),
             installed_app_id: None,
             membrane_proofs: HashMap::new(),
+            existing_cells: HashMap::new(),
             network_seed: None,
             ignore_genesis_failure: false,
         }));
