+++ crates/holochain/src/conductor/tests/mod.rs
@@ -2,5 +2,6 @@ mod app_info;
 mod cell_cloning;
 mod install_app_bundle;
 mod network_info;
+mod network_seed_regression;
 mod request_dna_def;
 mod signed_zome_call;
