+++ crates/holochain/src/sweettest/sweet_app_installation.rs
@@ -65,6 +65,7 @@ pub async fn get_install_app_payload_from_dnas(
         installed_app_id: Some(installed_app_id.into()),
         network_seed: None,
         membrane_proofs,
+        existing_cells: Default::default(),
         ignore_genesis_failure: false,
     }
 }
