+++ crates/holochain/tests/tests/test_utils/mod.rs
@@ -283,7 +283,8 @@ pub async fn register_and_install_dna_named(
         source: AppBundleSource::Bundle(bundle),
         installed_app_id: Some(name),
         network_seed: None,
-        membrane_proofs: std::collections::HashMap::new(),
+        membrane_proofs: Default::default(),
+        existing_cells: Default::default(),
         ignore_genesis_failure: false,
     };
     let request = AdminRequest::InstallApp(Box::new(payload));
