+++ crates/holochain_types/src/app/app_bundle.rs
@@ -63,6 +63,7 @@ impl AppBundle {
         dna_store: &impl DnaStore,
         agent: AgentPubKey,
         membrane_proofs: MemproofMap,
+        existing_cells: ExistingCellsMap,
     ) -> AppBundleResult<AppRoleResolution> {
         let AppManifestValidated { name: _, roles } = self.manifest().clone().validate()?;
         let bundle = Arc::new(self);
@@ -70,19 +71,28 @@ impl AppBundle {
             let bundle = bundle.clone();
             Ok((
                 role_name.clone(),
-                bundle.resolve_cell(dna_store, role_name, role).await?,
+                bundle
+                    .resolve_cell(dna_store, role_name, role, &existing_cells)
+                    .await?,
             ))
         });
-        let resolution = futures::future::join_all(tasks)
+
+        futures::future::join_all(tasks)
             .await
             .into_iter()
             .collect::<AppBundleResult<Vec<_>>>()?
             .into_iter()
             .fold(
-                AppRoleResolution::new(agent.clone()),
-                |mut resolution: AppRoleResolution, (role_name, op)| {
-                    match op {
-                        CellProvisioningOp::CreateFromDnaFile(dna, clone_limit) => {
+                Ok(AppRoleResolution::new(agent.clone())),
+                |mut resolution: AppBundleResult<AppRoleResolution>, (role_name, op)| {
+                    match (&mut resolution, op) {
+                        // Errors remain unchanged
+                        (Err(_), _) => (),
+
+                        (
+                            Ok(resolution),
+                            CellProvisioningOp::CreateFromDnaFile(dna, clone_limit),
+                        ) => {
                             let agent = resolution.agent.clone();
                             let dna_hash = dna.dna_hash().clone();
                             let cell_id = CellId::new(dna_hash, agent);
@@ -93,11 +103,12 @@ impl AppBundle {
                             resolution.role_assignments.push((role_name, role));
                         }
 
-                        CellProvisioningOp::Existing(cell_id, clone_limit) => {
+                        (Ok(resolution), CellProvisioningOp::Existing(cell_id, clone_limit)) => {
                             let role = AppRoleAssignment::new(cell_id, true, clone_limit);
                             resolution.role_assignments.push((role_name, role));
                         }
-                        CellProvisioningOp::ProvisionOnly(dna, clone_limit) => {
+
+                        (Ok(resolution), CellProvisioningOp::ProvisionOnly(dna, clone_limit)) => {
                             let agent = resolution.agent.clone();
                             let dna_hash = dna.dna_hash().clone();
                             let cell_id = CellId::new(dna_hash, agent);
@@ -110,22 +121,26 @@ impl AppBundle {
                                 AppRoleAssignment::new(cell_id, false, clone_limit),
                             ));
                         }
-                        other @ (CellProvisioningOp::HashMismatch(_, _)
-                        | CellProvisioningOp::Conflict(_)) => {
+
+                        (
+                            Ok(_),
+                            other @ (CellProvisioningOp::HashMismatch(_, _)
+                            | CellProvisioningOp::Unresolvable(_)),
+                        ) => {
                             tracing::error!(
                                 "Encountered unexpected CellProvisioningOp: {:?}",
                                 other
                             );
-                            unimplemented!()
+                            resolution = Err(AppBundleError::CellResolutionFailure(
+                                role_name,
+                                format!("{other:?}"),
+                            ));
                         }
                     }
 
                     resolution
                 },
-            );
-
-        // let resolution = cells.into_iter();
-        Ok(resolution)
+            )
     }
 
     async fn resolve_cell(
@@ -133,6 +148,7 @@ impl AppBundle {
         dna_store: &impl DnaStore,
         role_name: RoleName,
         role: AppRoleManifestValidated,
+        existing_cells: &ExistingCellsMap,
     ) -> AppBundleResult<CellProvisioningOp> {
         Ok(match role {
             AppRoleManifestValidated::Create {
@@ -155,40 +171,19 @@ impl AppBundle {
             }
 
             AppRoleManifestValidated::UseExisting {
-                installed_hash,
+                compatible_hash,
                 clone_limit,
                 deferred: _,
-            } => self.resolve_cell_existing(&installed_hash, clone_limit),
-            AppRoleManifestValidated::CreateIfNotExists {
-                location,
-                installed_hash,
-                clone_limit,
-                modifiers,
-                deferred: _,
-            } => match self.resolve_cell_existing(&installed_hash, clone_limit) {
-                op @ CellProvisioningOp::Existing(_, _) => op,
-                CellProvisioningOp::HashMismatch(_, _) => {
-                    let dna = self
-                        .resolve_dna(
-                            role_name,
-                            dna_store,
-                            &location,
-                            Some(&installed_hash),
-                            modifiers,
-                        )
-                        .await?;
-                    CellProvisioningOp::CreateFromDnaFile(dna, clone_limit)
+            } => {
+                if let Some(cell_id) = existing_cells.get(&role_name) {
+                    CellProvisioningOp::Existing(cell_id.clone(), clone_limit)
+                } else {
+                    CellProvisioningOp::Unresolvable(format!(
+                        "No existing cell was specified for the role with DNA {compatible_hash}"
+                    ))
                 }
-                CellProvisioningOp::Conflict(_) => {
-                    unimplemented!("conflicts are not handled, or even possible yet")
-                }
-                CellProvisioningOp::CreateFromDnaFile(_, _) => {
-                    unreachable!("resolve_cell_existing will never return a Create op")
-                }
-                CellProvisioningOp::ProvisionOnly(_, _) => {
-                    unreachable!("resolve_cell_existing will never return a ProvisionOnly")
-                }
-            },
+            }
+
             AppRoleManifestValidated::CloneOnly {
                 clone_limit,
                 location,
@@ -240,14 +235,6 @@ impl AppBundle {
         Ok(dna_file)
     }
 
-    fn resolve_cell_existing(
-        &self,
-        _version: &DnaHashB64,
-        _clone_limit: u32,
-    ) -> CellProvisioningOp {
-        unimplemented!("Reusing existing cells is not yet implemented")
-    }
-
     async fn resolve_location(
         &self,
         location: &mr_bundle::Location,
@@ -258,12 +245,6 @@ impl AppBundle {
         let (dna_file, original_hash) = dna_bundle.into_dna_file(modifiers).await?;
         Ok((dna_file, original_hash))
     }
-}
-
-/// This function is called in places where it will be necessary to rework that
-/// area after use_existing has been implemented
-pub fn we_must_remember_to_rework_cell_panic_handling_after_implementing_use_existing_cell_resolution(
-) {
 }
 
 /// The answer to the question:
@@ -330,10 +311,6 @@ pub enum CellProvisioningOp {
     ProvisionOnly(DnaFile, u32),
     /// The specified installed_hash does not match the actual hash of the DNA selected for provisioning. Expected: {0}, Actual: {1}
     HashMismatch(DnaHash, DnaHash),
-    /// Ambiguous result, needs manual resolution; can't provision (should this be an Err?)
-    Conflict(CellProvisioningConflict),
+    /// Ambiguous result, needs more information; can't provision (should this be an Err?)
+    Unresolvable(String),
 }
-
-/// Uninhabitable placeholder
-#[derive(Debug)]
-pub enum CellProvisioningConflict {}
