+++ crates/holochain_types/src/app/app_bundle/tests.rs
@@ -61,6 +61,7 @@ async fn provisioning_1_create() {
             &std::collections::HashMap::new(),
             agent.clone(),
             Default::default(),
+            Default::default(),
         )
         .await
         .unwrap();
